requirements.txt
pandas==1.2.4
imutils==0.5.4
matplotlib==3.4.2
numpy==1.19.5
openpyxl==3.0.7
Pillow==8.2.0
pydicom==2.1.2
scikit-image==0.18.1
scipy==1.6.3
tqdm==4.61.0
Keras==2.4.3
torch==1.8.1
PyYAML==5.4.1
opencv-python==4.5.2.52
torchvision==0.9.1